+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Validation Framework Version History
 
+**Version 1.20**
+
+- Updated Groovy library from version 2.4.12 (indy) to version 2.5.1 (indy).
+- Updated Apache commons-lang library from version 3.6 to version 3.7.
+
 **Version 1.19**
 
 - Changed context function getEodSchemaId() to use 'behavior' as an input.